product_service/pyproject.toml
[tool.poetry]
name = "product-service"
version = "0.1.0"
description = ""
authors = ["artorias <kerelkobarabash@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
mongorepo = "^0.1.3"
pydantic-settings = "^2.3.4"
punq = "^0.7.0"
aio-pika = "^9.4.2"
orjson = "^3.10.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
